# openapi/openapi.yaml
openapi: 3.1.2
info:
  title: ocservia API
  version: 1.0.0
  description: Public control-plane API contract.
  license:
    name: MIT OR Apache-2.0
    identifier: MIT OR Apache-2.0
servers:
  - url: /api/v1
security:
  - oidc: []
  - bearerAuth: []
...
paths:
...
  /role-bindings:
    post:
      operationId: createRoleBinding
      summary: Bind one baseline role to a workspace or resource scope
      tags: [platform]
      parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RoleBindingRequest" } } }
      responses:
        "201":
          description: Role binding created.
          content: { application/json: { schema: { $ref: "#/components/schemas/RoleBinding" } } }
        "403": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    oidc:
      type: apiKey
      in: cookie
      name: __Host-ocservia_session
    bearerAuth:
      type: http
      scheme: bearer
      description: Secure, HttpOnly, SameSite server-side session established by OIDC Authorization Code plus PKCE S256.
  parameters:
...
    WorkspaceId:
      name: X-Workspace-ID
      in: header
      description: Required when the principal can access more than one workspace; otherwise the sole authorized workspace is inferred.
      required: false
      schema: { $ref: "#/components/schemas/UuidV7" }
...
  responses:
    Problem:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    UuidV7:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      example: 0193c1b2-7d8a-7f01-8a2d-1e5b4f302901
...
    RoleBindingRequest:
      type: object
      additionalProperties: false
      required: [identity_id, workspace_id, role, resource_type, reason]
      properties:
        identity_id: { $ref: "#/components/schemas/UuidV7" }
        workspace_id: { $ref: "#/components/schemas/UuidV7" }
        role: { type: string, enum: [Viewer, Operator, UserManager, ConfigManager, Auditor, SecurityAdmin, PlatformAdmin] }
        resource_type: { type: string, enum: [workspace, node, resource, secret_ref, certificate, config_plan, batch_operation, role_binding] }
        resource_id: { $ref: "#/components/schemas/UuidV7" }
        reason: { type: string, minLength: 1, maxLength: 512 }
        approval_id:
          allOf: [{ $ref: "#/components/schemas/UuidV7" }]
          description: Required for SecurityAdmin and PlatformAdmin grants.
    RoleBinding:
      type: object
      additionalProperties: false
      required: [id]
      properties:
        id: { $ref: "#/components/schemas/UuidV7" }
...
    Problem:
      type: object
      additionalProperties: true
      description: RFC 9457 Problem Details response.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
          minLength: 1
          maxLength: 256
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          maxLength: 4096
        instance:
          type: string
          format: uri-reference
        trace_id:
          type: string
          pattern: "^[0-9a-f]{32}$"